+++ common/src/main/kotlin/farm/rosehearth/compatemon/CompatemonKotlin.kt
@@ -17,6 +17,8 @@ import farm.rosehearth.compatemon.utils.CompatemonDataKeys.COMPAT_SCALE_WEIGHT
 import farm.rosehearth.compatemon.utils.CompatemonDataKeys.MOD_ID_COMPATEMON
 import farm.rosehearth.compatemon.utils.CompatemonDataKeys.MOD_ID_PEHKUI
 import farm.rosehearth.compatemon.utils.CompatemonDataKeys.MOD_ID_APOTHEOSIS
+import farm.rosehearth.compatemon.utils.CompatemonDataKeys.APOTH_BOSS
+import farm.rosehearth.compatemon.utils.CompatemonDataKeys.APOTH_RARITY
 import farm.rosehearth.compatemon.modules.pehkui.util.CompatemonScaleUtils
 import net.minecraft.ChatFormatting
 import net.minecraft.network.chat.*
@@ -58,9 +60,9 @@ object CompatemonKotlin {
         // Occurs when the Pokemon's NBT data is loaded from the world.
         POKEMON_ENTITY_LOAD.subscribe{ event ->
             if(Compatemon.ShouldLoadMod(MOD_ID_PEHKUI)) {
-                //var sizeScale: Float = CompatemonScaleUtils.getScale(event.pokemonEntity.pokemon, "$MOD_ID_PEHKUI:$COMPAT_SCALE_SIZE")
-                //var weightScale: Float = CompatemonScaleUtils.getScale(event.pokemonEntity.pokemon, "$MOD_ID_COMPATEMON:$COMPAT_SCALE_WEIGHT")
-               // CompatemonScaleUtils.setScale(event.pokemonEntity, ScaleTypes.BASE, "$MOD_ID_PEHKUI:$COMPAT_SCALE_SIZE", sizeScale)
+                var sizeScale: Float = CompatemonScaleUtils.getScale(event.pokemonEntity.pokemon, "$MOD_ID_PEHKUI:$COMPAT_SCALE_SIZE")
+               // var weightScale: Float = CompatemonScaleUtils.getScale(event.pokemonEntity.pokemon, "$MOD_ID_COMPATEMON:$COMPAT_SCALE_WEIGHT")
+                CompatemonScaleUtils.setScale(event.pokemonEntity, ScaleTypes.BASE, "$MOD_ID_PEHKUI:$COMPAT_SCALE_SIZE", sizeScale)
             }
         }
         // Occurs when the Pokemon is Sent Out of its pokeball
@@ -85,14 +87,15 @@ object CompatemonKotlin {
             }
         }
 
+        // To update the nickname color whenever the name changes
         POKEMON_NICKNAMED.subscribe{event ->
            // event.pokemon.nickname = event.nickname?.withStyle(Style.EMPTY.withColor(ChatFormatting.GOLD))
             if(Compatemon.ShouldLoadMod(MOD_ID_APOTHEOSIS)){
-                var isBoss = event.pokemon.persistentData.getCompound(MOD_ID_COMPATEMON).contains("apoth.boss");
+                var isBoss = event.pokemon.persistentData.getCompound(MOD_ID_COMPATEMON).contains(APOTH_BOSS);
                 if(isBoss){
-                    var rarityKey = event.pokemon.persistentData.getCompound(MOD_ID_COMPATEMON).getString("apoth.rarity.color")
+                    var rarityKey = event.pokemon.persistentData.getCompound(MOD_ID_COMPATEMON).getString("$APOTH_RARITY.color")
                     event.pokemon.nickname = event.pokemon.nickname?.withStyle(Style.EMPTY.withColor(TextColor.parseColor(rarityKey)))
-                    Compatemon.LOGGER.debug(event.pokemon.nickname.toString())
+                    LOGGER.debug(event.pokemon.nickname.toString())
                 }
             }
         }