+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-02-09
+
+### Added
+- Apotheosis Pokemon Bosses Spawn Properly - kind of. Some pokemon spawn as bosses but aren't supposed to.
+- > Pokemon can drop affix items and gems based on config.
+- > Boss sizes scale based on config. Unsure if this works with normal bosses.
+- Sophisticated Core functionality to make Controller Search Range based on config - UNTESTED
+- 
+### Fixed
+- Apotheosis Issues with pokemon bosses not spawning AS bosses or with items
+
+### Changed
+- Updated TODO file! YAY!
 
 ## [0.1.3] - 2024-02-09
 
@@ -13,12 +26,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Everything!
 - Pehkui Integration for Fabric and Forge
 - Apotheosis Integration for Forge
-- Pokemon Music Disks: “ENLS’s Pre-Looped Music Library” extending Quark's Ambient Disks
-- 
+- Pokemon Music Disks: “ENLS’s Pre-Looped Music Library” extending Quark's Ambient Disks (ish)
+-
 
 ### Fixed
 
-- 
+
 
 ### Changed